feat: add product image upload and pdf invoice generating for orders

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require("express-validator")
+const fileHelper = require("../util/file")
 
 const Product = require("../models/product")
 
@@ -14,7 +15,23 @@ exports.getAddProduct = (req, res, next) => {
 }
 
 exports.postAddProduct = (req, res, next) => {
-    const { title, price, description, imageUrl } = req.body
+    const { title, price, description } = req.body
+    const image = req.file
+
+    if (!image) {
+        return res.status(422).render("admin/edit-product", {
+            pageTitle: "Add Product",
+            path: "/admin/add-product",
+            editing: false,
+            hasError: true,
+            product: { title, price, description },
+            errorMessage: "Attached file is not an image.",
+            validationErrors: [],
+        })
+    }
+
+    const imageUrl = image.path
+
     const product = new Product({
         title,
         price,
@@ -74,7 +91,8 @@ exports.getEditProduct = (req, res, next) => {
 }
 
 exports.postEditProduct = (req, res, next) => {
-    const { productId, title, imageUrl, price, description } = req.body
+    const { productId, title, price, description } = req.body
+    const image = req.file
     const errors = validationResult(req)
 
     if (!errors.isEmpty()) {
@@ -83,7 +101,7 @@ exports.postEditProduct = (req, res, next) => {
             path: "/admin/edit-product",
             editing: true,
             hasError: true,
-            product: { title, price, description, imageUrl, _id: productId },
+            product: { title, price, description, _id: productId },
             errorMessage: errors.array()[0].msg,
             validationErrors: errors.array(),
         })
@@ -95,9 +113,12 @@ exports.postEditProduct = (req, res, next) => {
                 return res.redirect("/")
             }
             product.title = title
-            product.imageUrl = imageUrl
             product.price = price
             product.description = description
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl)
+                product.imageUrl = image.path
+            }
             return product.save().then((result) => {
                 console.log("UPDATED PRODUCT")
                 res.redirect("/admin/products")
@@ -112,8 +133,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId
-
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then((product) => {
+            if (!product) {
+                return next(new Error("Product not found."))
+            }
+            fileHelper.deleteFile(product.imageUrl)
+            return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        })
         .then(() => {
             res.redirect("/admin/products")
         })
